external api tests pass for game apis
