+++ src/main/resources/application.properties
@@ -38,5 +38,6 @@ spring.profiles.active=dev
 # 4. Generate access token by making POST request to:
 #    https://id.twitch.tv/oauth2/token?client_id=YOUR_CLIENT_ID&client_secret=YOUR_CLIENT_SECRET&grant_type=client_credentials
 
-igdb.client.id=your-twitch-client-id-here
-igdb.access.token=your-igdb-access-token-here
+igdb.client.id=egczzym1ranmhiqsv4f5s5tc39qibf
+igdb.access.token=hz0brxgv2i6ptvcn4xfwp3ge5r6nv4
+#50x0hvcmf8o3x5m3bd5re8a0gurgzz
